+++ src/main/java/org/folio/holdingsiq/model/ResourceId.java
@@ -1,7 +1,7 @@
 package org.folio.holdingsiq.model;
 
 public record ResourceId(
-  long providerIdPart,
-  long packageIdPart,
-  long titleIdPart
+  int providerIdPart,
+  int packageIdPart,
+  int titleIdPart
 ) { }
